imrpoved editor curosr

+++ scripts/services/EditorService.gd
@@ -9,14 +9,15 @@ class_name EditorService
 var editor_ui: EditorUI
 var player_menu: PlayerMenu
 static var cursor: HexCursor
-static var selected_type: String = "grass"
+static var selected_type: String = "hex"
+static var selected_subtype: String = "grass"
 static var selected_settlement: Settlement
 
 func _ready() -> void:
 	GameStateService.editor_service = self
 	cursor = hex_scene.instantiate()
 	GameStateService.world_manager.hexes.add_child(cursor)
-	cursor.set_cursor_type(selected_type)
+	cursor.set_cursor_type(selected_type, selected_subtype)
 	editor_ui = editor_ui_scene.instantiate()
 	$"../CanvasLayer".add_child(editor_ui)
 
@@ -27,30 +28,33 @@ func move_editor_cursor(move_to_hex: Hex) -> void:
 	cursor.q = move_to_hex.q
 	cursor.r = move_to_hex.r
 	
-func _on_editor_ui_type_changed(type: String) -> void:
-	cursor.set_cursor_type(type)
+func _on_editor_ui_type_changed(type: String, sub_type: String = "") -> void:
+	cursor.set_cursor_type(type, sub_type)
 	selected_type = type
+	selected_subtype = sub_type
 
 func hex_clicked(hex: Hex, event: InputEvent):
-	if selected_type == "settlement":
+	if selected_type == "object" && selected_subtype == "settlement":
 		GameStateService.data_service.add_settlement(hex, null)
-	elif selected_type == "select":
+	elif selected_type == "action" && selected_subtype == "select":
 		selected_settlement = GameStateService.data_service.get_settlement_by_hex(hex)
 		if selected_settlement != null:
 			editor_ui.set_settlement_info(selected_settlement)
-	elif selected_type == "river":
+	elif selected_type == "object" && selected_subtype == "river":
 		var sextant_clicked: Hex.side_flag = hex.get_sextant_clicked(event)
 		if GameStateService.data_service.add_river(hex, sextant_clicked):
 			GameStateService.world_manager.add_river(hex, sextant_clicked)
-	else:
-		GameStateService.data_service.update_hex_type(hex, selected_type)
+	elif selected_type == "unit":
+		pass
+	elif selected_type == "hex":
+		GameStateService.data_service.update_hex_type(hex, selected_subtype)
 		
 
 func hex_alt_clicked(hex: Hex, event: InputEvent):
-	if selected_type == "settlement":
+	if selected_type == "object" && selected_subtype == "settlement":
 		GameStateService.data_service.remove_settlement(hex)
 		GameStateService.world_manager.refresh_hex_settlement(GameStateService.data_service.world_dict, hex)
-	elif selected_type == "river":
+	elif selected_type == "object" && selected_subtype == "river":
 		var sextant_clicked: Hex.side_flag = hex.get_sextant_clicked(event)
 		GameStateService.data_service.remove_river(hex, sextant_clicked)
 		GameStateService.world_manager.refresh_hex_rivers(GameStateService.data_service.world_dict, hex)
